Change to use redux-toolkit Enact-DCO-1.0-Signed-off-by: SJ RO (sj.ro@lge.com)
