+++ samples/event-logger/src/store/slices/eventCapturingOnSlice.js
@@ -1,0 +1,14 @@
+import {createSlice} from '@reduxjs/toolkit';
+
+export const eventCapturingOnSlice = createSlice( {
+	name: 'eventCapturingOnReducer',
+	initialState: false,
+	reducers: {
+		setEventCapturing: (state, action) => {
+			return (state = action.payload);
+		}
+	}
+});
+
+export const {setEventCapturing} = eventCapturingOnSlice.actions;
+export default eventCapturingOnSlice;
