+++ samples/event-logger/src/store/slices/syntheticEventOnSlice.js
@@ -1,0 +1,14 @@
+import {createSlice} from '@reduxjs/toolkit';
+
+export const syntheticEventOnSlice = createSlice( {
+	name: 'syntheticEventOnReducer',
+	initialState: false,
+	reducers: {
+		isSyntheticEventOn: (state, action) => {
+			return (state = action.payload);
+		}
+	}
+});
+
+export const {isSyntheticEventOn} = syntheticEventOnSlice.actions;
+export default syntheticEventOnSlice;
